feat: update Tile component #150
