+++ src/Tile.tsx
@@ -1,10 +1,10 @@
-import React, { memo, forwardRef, type ReactNode, type CSSProperties } from "react";
-import { symToStr } from "tsafe/symToStr";
-import { assert } from "tsafe/assert";
+import React, { ComponentProps, CSSProperties, forwardRef, memo, ReactNode } from "react";
 import type { Equals } from "tsafe";
-import type { RegisteredLinkProps } from "./link";
-import { getLink } from "./link";
+import { assert } from "tsafe/assert";
+import { symToStr } from "tsafe/symToStr";
+
 import { fr } from "./fr";
+import { type RegisteredLinkProps, getLink } from "./link";
 import { cx } from "./tools/cx";
 import { useAnalyticsId } from "./tools/useAnalyticsId";
 
@@ -13,25 +13,64 @@ export type TileProps = {
     id?: string;
     className?: string;
     title: ReactNode;
-    linkProps: RegisteredLinkProps;
+    titleAs?: `h${2 | 3 | 4 | 5 | 6}`;
+    linkProps?: RegisteredLinkProps;
+    buttonProps?: ComponentProps<"button">;
+    downloadButton?: boolean;
     desc?: ReactNode;
+    detail?: ReactNode;
+    start?: ReactNode;
     imageUrl?: string;
     imageAlt?: string;
     imageWidth?: string | number;
     imageHeight?: string | number;
+    imageSvg?: boolean;
     grey?: boolean;
-
     /** make the whole tile clickable */
-    enlargeLink?: boolean;
+    enlargeLinkOrButton?: boolean;
     classes?: Partial<
-        Record<"root" | "title" | "link" | "body" | "desc" | "img" | "imgTag", string>
+        Record<
+            | "root"
+            | "content"
+            | "title"
+            | "header"
+            | "link"
+            | "button"
+            | "body"
+            | "desc"
+            | "detail"
+            | "start"
+            | "img"
+            | "imgTag",
+            string
+        >
     >;
-    /** Default false */
-    horizontal?: boolean;
+    orientation?: "horizontal" | "vertical";
+    small?: boolean;
+    noIcon?: boolean;
+    noBorder?: boolean;
+    noBackground?: boolean;
+    disabled?: boolean;
     style?: CSSProperties;
-};
+} & (TileNextProps.WithLink | TileNextProps.WithButton | TileNextProps.Unclickable);
 
-export namespace TileProps {}
+export namespace TileNextProps {
+    export type Unclickable = {
+        linkProps?: never;
+        buttonProps?: never;
+        enlargeLinkOrButton?: never;
+    };
+
+    export type WithLink = {
+        linkProps: RegisteredLinkProps;
+        buttonProps?: never;
+    };
+
+    export type WithButton = {
+        linkProps?: never;
+        buttonProps: ComponentProps<"button">;
+    };
+}
 
 /** @see <https://components.react-dsfr.codegouv.studio/?path=/docs/components-tile> */
 export const Tile = memo(
@@ -40,16 +79,27 @@ export const Tile = memo(
             id: id_props,
             className,
             title,
+            titleAs: HtmlTitleTag = "h3",
             linkProps,
+            buttonProps,
+            downloadButton,
             desc,
+            detail,
+            start,
             imageUrl,
             imageAlt,
             imageWidth,
             imageHeight,
-            horizontal = false,
+            imageSvg = true,
+            orientation = "vertical",
+            small = false,
+            noBorder = false,
+            noIcon = false,
+            noBackground = false,
             grey = false,
             classes = {},
-            enlargeLink = true,
+            enlargeLinkOrButton = true,
+            disabled = false,
             style,
             ...rest
         } = props;
@@ -63,15 +113,29 @@ export const Tile = memo(
             "explicitlyProvidedId": id_props
         });
 
+        if (!!disabled && linkProps !== undefined) {
+            linkProps.href = undefined;
+        }
+
         return (
             <div
                 id={id}
                 className={cx(
                     fr.cx(
                         "fr-tile",
-                        enlargeLink && "fr-enlarge-link",
-                        horizontal && "fr-tile--horizontal",
-                        grey && "fr-tile--grey"
+                        enlargeLinkOrButton &&
+                            (linkProps
+                                ? "fr-enlarge-link"
+                                : buttonProps
+                                ? "fr-enlarge-button"
+                                : null),
+                        orientation && `fr-tile--${orientation}`,
+                        noIcon && "fr-tile--no-icon",
+                        noBorder && "fr-tile--no-border",
+                        noBackground && "fr-tile--no-background",
+                        grey && "fr-tile--grey",
+                        small && "fr-tile--sm",
+                        buttonProps && downloadButton && "fr-tile--download"
                     ),
                     classes.root,
                     className
@@ -81,30 +145,89 @@ export const Tile = memo(
                 {...rest}
             >
                 <div className={cx(fr.cx("fr-tile__body"), classes.body)}>
-                    <h3 className={cx(fr.cx("fr-tile__title"), classes.title)}>
-                        <Link {...linkProps} className={cx(classes.link, linkProps.className)}>
-                            {title}
-                        </Link>
-                    </h3>
-                    <p className={cx(fr.cx("fr-tile__desc"), classes.desc)}>{desc}</p>
-                </div>
-                {(imageUrl !== undefined && imageUrl.length && (
-                    <div className={cx(fr.cx("fr-tile__img"), classes.img)}>
-                        <img
-                            className={cx(fr.cx("fr-responsive-img"), classes.imgTag)}
-                            src={imageUrl}
-                            alt={imageAlt}
-                            width={imageWidth}
-                            height={imageHeight}
-                        />
+                    <div className={cx(fr.cx("fr-tile__content"), classes.content)}>
+                        <HtmlTitleTag className={cx(fr.cx("fr-tile__title"), classes.title)}>
+                            {linkProps !== undefined ? (
+                                <Link
+                                    {...linkProps}
+                                    className={cx(classes.link, linkProps.className)}
+                                    aria-disabled={!!disabled}
+                                >
+                                    {title}
+                                </Link>
+                            ) : buttonProps !== undefined ? (
+                                <button
+                                    {...buttonProps}
+                                    className={cx(classes.button, buttonProps.className)}
+                                    disabled={!!disabled}
+                                >
+                                    {title}
+                                </button>
+                            ) : (
+                                title
+                            )}
+                        </HtmlTitleTag>
+
+                        {desc !== undefined && (
+                            <p className={cx(fr.cx("fr-tile__desc"), classes.desc)}>{desc}</p>
+                        )}
+                        {detail !== undefined && (
+                            <p className={cx(fr.cx("fr-tile__detail"), classes.detail)}>{detail}</p>
+                        )}
+                        {start !== undefined && (
+                            <div className={cx(fr.cx("fr-tile__start"), classes.start)}>
+                                {start}
+                            </div>
+                        )}
                     </div>
-                )) ||
-                    null}
+                </div>
+
+                {imageUrl !== undefined && imageUrl.length && (
+                    <div className={cx(fr.cx("fr-tile__header"), classes.header)}>
+                        {imageSvg ? (
+                            <div className={cx(fr.cx("fr-tile__pictogram"), classes.img)}>
+                                <svg
+                                    aria-hidden={true}
+                                    className={fr.cx("fr-artwork")}
+                                    viewBox="0 0 80 80"
+                                    width="80px"
+                                    height="80px"
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    xmlnsXlink="http://www.w3.org/1999/xlink"
+                                >
+                                    {(
+                                        [
+                                            "artwork-decorative",
+                                            "artwork-minor",
+                                            "artwork-major"
+                                        ] as const
+                                    ).map(label => (
+                                        <use
+                                            key={label}
+                                            className={fr.cx(`fr-${label}`)}
+                                            xlinkHref={`${imageUrl}#${label}`}
+                                        />
+                                    ))}
+                                </svg>
+                            </div>
+                        ) : (
+                            <div className={cx(fr.cx("fr-tile__img"), classes.img)}>
+                                <img
+                                    className={cx(fr.cx("fr-responsive-img"), classes.imgTag)}
+                                    src={imageUrl}
+                                    alt={imageAlt}
+                                    width={imageWidth}
+                                    height={imageHeight}
+                                />
+                            </div>
+                        )}
+                    </div>
+                )}
             </div>
         );
     })
 );
 
-Tile.displayName = symToStr({ Tile });
+Tile.displayName = symToStr({ TileNext: Tile });
 
 export default Tile;
