--- src/TileNext.tsx
@@ -1,229 +1,0 @@
-import React, { ComponentProps, CSSProperties, forwardRef, memo, ReactNode } from "react";
-import { type Equals, assert } from "tsafe";
-import { symToStr } from "tsafe/symToStr";
-
-import { fr } from "./fr";
-import { type RegisteredLinkProps, getLink } from "./link";
-import { cx } from "./tools/cx";
-import { useAnalyticsId } from "./tools/useAnalyticsId";
-
-export type TileNextProps = {
-    id?: string;
-    className?: string;
-    title: ReactNode;
-    titleAs?: `h${2 | 3 | 4 | 5 | 6}`;
-    linkProps?: RegisteredLinkProps;
-    buttonProps?: ComponentProps<"button">;
-    downloadButton?: boolean;
-    desc?: ReactNode;
-    detail?: ReactNode;
-    start?: ReactNode;
-    imageUrl?: string;
-    imageAlt?: string;
-    imageWidth?: string | number;
-    imageHeight?: string | number;
-    imageSvg?: boolean;
-    grey?: boolean;
-    /** make the whole tile clickable */
-    enlargeLinkOrButton?: boolean;
-    classes?: Partial<
-        Record<
-            | "root"
-            | "content"
-            | "title"
-            | "header"
-            | "link"
-            | "button"
-            | "body"
-            | "desc"
-            | "detail"
-            | "start"
-            | "img"
-            | "imgTag",
-            string
-        >
-    >;
-    orientation?: "horizontal" | "vertical";
-    small?: boolean;
-    noIcon?: boolean;
-    noBorder?: boolean;
-    noBackground?: boolean;
-    disabled?: boolean;
-    style?: CSSProperties;
-} & (TileNextProps.WithLink | TileNextProps.WithButton | TileNextProps.Unclickable);
-
-export namespace TileNextProps {
-    export type Unclickable = {
-        linkProps?: never;
-        buttonProps?: never;
-        enlargeLinkOrButton?: never;
-    };
-
-    export type WithLink = {
-        linkProps: RegisteredLinkProps;
-        buttonProps?: never;
-    };
-
-    export type WithButton = {
-        linkProps?: never;
-        buttonProps: ComponentProps<"button">;
-    };
-}
-
-/** @see <https://components.react-dsfr.codegouv.studio/?path=/docs/components-tilenext--default> */
-export const TileNext = memo(
-    forwardRef<HTMLDivElement, TileNextProps>((props, ref) => {
-        const {
-            id: id_props,
-            className,
-            title,
-            titleAs: HtmlTitleTag = "h3",
-            linkProps,
-            buttonProps,
-            downloadButton,
-            desc,
-            detail,
-            start,
-            imageUrl,
-            imageAlt,
-            imageWidth,
-            imageHeight,
-            imageSvg = true,
-            orientation = "vertical",
-            small = false,
-            noBorder = false,
-            noIcon = false,
-            noBackground = false,
-            grey = false,
-            classes = {},
-            enlargeLinkOrButton = true,
-            disabled = false,
-            style,
-            ...rest
-        } = props;
-
-        assert<Equals<keyof typeof rest, never>>();
-
-        const { Link } = getLink();
-
-        const id = useAnalyticsId({
-            "defaultIdPrefix": "fr-tile",
-            "explicitlyProvidedId": id_props
-        });
-
-        if (!!disabled && linkProps !== undefined) {
-            linkProps.href = undefined;
-        }
-
-        return (
-            <div
-                id={id}
-                className={cx(
-                    fr.cx(
-                        "fr-tile",
-                        enlargeLinkOrButton &&
-                            (linkProps
-                                ? "fr-enlarge-link"
-                                : buttonProps
-                                ? "fr-enlarge-button"
-                                : null),
-                        orientation && `fr-tile--${orientation}`,
-                        noIcon && "fr-tile--no-icon",
-                        noBorder && "fr-tile--no-border",
-                        noBackground && "fr-tile--no-background",
-                        grey && "fr-tile--grey",
-                        small && "fr-tile--sm",
-                        buttonProps && downloadButton && "fr-tile--download"
-                    ),
-                    classes.root,
-                    className
-                )}
-                ref={ref}
-                style={style}
-                {...rest}
-            >
-                <div className={cx(fr.cx("fr-tile__body"), classes.body)}>
-                    <div className={cx(fr.cx("fr-tile__content"), classes.content)}>
-                        <HtmlTitleTag className={cx(fr.cx("fr-tile__title"), classes.title)}>
-                            {linkProps ? (
-                                <Link
-                                    {...linkProps}
-                                    className={cx(classes.link, linkProps.className)}
-                                    aria-disabled={!!disabled}
-                                >
-                                    {title}
-                                </Link>
-                            ) : buttonProps ? (
-                                <button
-                                    {...buttonProps}
-                                    className={cx(classes.button, buttonProps.className)}
-                                    disabled={!!disabled}
-                                >
-                                    {title}
-                                </button>
-                            ) : (
-                                title
-                            )}
-                        </HtmlTitleTag>
-
-                        {desc && <p className={cx(fr.cx("fr-tile__desc"), classes.desc)}>{desc}</p>}
-                        {detail && (
-                            <p className={cx(fr.cx("fr-tile__detail"), classes.detail)}>{detail}</p>
-                        )}
-                        {start && (
-                            <div className={cx(fr.cx("fr-tile__start"), classes.start)}>
-                                {start}
-                            </div>
-                        )}
-                    </div>
-                </div>
-
-                {imageUrl !== undefined && imageUrl.length && (
-                    <div className={cx(fr.cx("fr-tile__header"), classes.header)}>
-                        {imageSvg ? (
-                            <div className={fr.cx("fr-tile__pictogram")}>
-                                <svg
-                                    aria-hidden={true}
-                                    className={fr.cx("fr-artwork")}
-                                    viewBox="0 0 80 80"
-                                    width="80px"
-                                    height="80px"
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    xmlnsXlink="http://www.w3.org/1999/xlink"
-                                >
-                                    {(
-                                        [
-                                            "artwork-decorative",
-                                            "artwork-minor",
-                                            "artwork-major"
-                                        ] as const
-                                    ).map(label => (
-                                        <use
-                                            key={label}
-                                            className={fr.cx(`fr-${label}`)}
-                                            xlinkHref={`${imageUrl}#${label}`}
-                                        />
-                                    ))}
-                                </svg>
-                            </div>
-                        ) : (
-                            <div className={cx(fr.cx("fr-tile__img"), classes.img)}>
-                                <img
-                                    className={cx(fr.cx("fr-responsive-img"), classes.imgTag)}
-                                    src={imageUrl}
-                                    alt={imageAlt}
-                                    width={imageWidth}
-                                    height={imageHeight}
-                                />
-                            </div>
-                        )}
-                    </div>
-                )}
-            </div>
-        );
-    })
-);
-
-TileNext.displayName = symToStr({ TileNext });
-
-export default TileNext;
